+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Road map
 
+- [ ] Check if file is a source file before transpiling
 - [ ] Move strings to resource file
 - [ ] Minification of the generated .js